+++ src/gs/gsEdge.h
@@ -25,7 +25,7 @@ namespace gs
         const int                       id;
         const shared_ptr<gs::Vertex>    v0;
         const shared_ptr<gs::Vertex>    v1;
-		gs::Vec3f                       vec;
+		gs::Vec3d                       vec;
 
     public:
         bool                            AddTile(const shared_ptr<gs::Tile>& newTile);
